Change Env::SchedClosureAfter and Env::SleepMicroseconds to take micros as an int64. Change posix env to sleep using nanosleep and retry on interrupt signal. Change: 126745229

+++ tensorflow_serving/test_util/fake_clock_env.cc
@@ -72,7 +72,7 @@ uint64 FakeClockEnv::NowMicros() {
   }
 }
 
-void FakeClockEnv::SleepForMicroseconds(int micros) {
+void FakeClockEnv::SleepForMicroseconds(int64 micros) {
   if (micros == 0) {
     return;
   }
